+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added minimum balance checks and precision validation for withdrawal amounts [DV-3217]
 
 ## [0.9.5] - 2025-09-15
 - Fix walidation withdrawal exchnage [DV-3362]